doxybuild + cache

+++ ci/travis-install-doxygen.sh
@@ -3,6 +3,12 @@
 
 set -ex
 
+FILE=$HOME/local/doxygen    
+if [ -f $FILE ]; then
+  echo "File $FILE exists. Skipping."
+  exit 0
+fi
+
 wget -O doxygen.tar.gz https://github.com/doxygen/doxygen/archive/Release_1_8_14.tar.gz
 tar -xvf doxygen.tar.gz
 rm doxygen.tar.gz
@@ -11,6 +17,6 @@ mkdir build
 cd build
 cmake -G "Unix Makefiles" ..
 make
-make install
+mkdir -p $HOME/local
+mv bin/* $HOME/local
 cd ..
-rm -rf doxygen-Release_1_8_14
